+++ rest_framework/utils/model_meta.py
@@ -12,7 +12,7 @@ from django.core.exceptions import ImproperlyConfigured
 from django.db import models
 from django.utils import six
 
-from rest_framework.compat import OrderedDict
+from rest_framework.compat import OrderedDict, get_model
 
 FieldInfo = namedtuple('FieldResult', [
     'pk',  # Model field instance
@@ -45,7 +45,7 @@ def _resolve_model(obj):
     """
     if isinstance(obj, six.string_types) and len(obj.split('.')) == 2:
         app_name, model_name = obj.split('.')
-        resolved_model = models.get_model(app_name, model_name)
+        resolved_model = get_model(app_name, model_name)
         if resolved_model is None:
             msg = "Django did not return a model for {0}.{1}"
             raise ImproperlyConfigured(msg.format(app_name, model_name))