+++ docs/topics/release-notes.md
@@ -38,6 +38,14 @@ You can determine your currently installed version using `pip freeze`:
 
 ---
 
+## 3.3.x series
+
+### 3.3.0
+
+**Date**: NOT YET RELEASED
+
+* Removed support for Django Version 1.5 ([#3421][gh3421])
+
 ## 3.2.x series
 
 ### 3.2.4
@@ -533,3 +541,6 @@ For older release notes, [please see the version 2.x documentation][old-release-
 [gh3361]: https://github.com/tomchristie/django-rest-framework/issues/3361
 [gh3364]: https://github.com/tomchristie/django-rest-framework/issues/3364
 [gh3415]: https://github.com/tomchristie/django-rest-framework/issues/3415
+
+<!-- 3.3.0 -->
+[gh3421]: https://github.com/tomchristie/django-rest-framework/pulls/3421